Rename: move public icons -> assets
