+++ app/_component/common/header/_component/Close.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useRouter } from 'next/navigation'
-import CloseIcon from '@/public/close.svg'
+import CloseIcon from '@/public/assets/close.svg'
 
 export default function Close() {
   const router = useRouter()
