+++ app/_component/common/header/_component/Heart.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useRouter } from 'next/navigation'
-import HeartIcon from '@/public/heart.svg'
+import HeartIcon from '@/public/assets/heart.svg'
 
 export default function Heart() {
   const router = useRouter()
