+++ app/_component/common/header/_component/Search.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useRouter } from 'next/navigation'
-import SearchIcon from '@/public/search.svg'
+import SearchIcon from '@/public/assets/search.svg'
 
 export default function Search() {
   const router = useRouter()
