+++ app/_component/common/header/header.module.css
@@ -1,7 +1,7 @@
 .container {
   width: 100%;
   height: 48px;
-  padding: 12px 0;
+  padding-bottom: 12px;
   display: flex;
   align-items: center;
   flex-direction: row;
@@ -17,7 +17,6 @@
 .title {
   text-align: center;
   margin: 0 auto;
-
   font-family: 'Pretendard Variable';
   font-size: 17px;
   font-style: normal;
